Class type OCanren.reified

Reification result

method is_open : bool

Returns true if the term has any free logic variable inside

method prj : 'a

Gets the answer as regular term. Raises exception Not_a_value when the answer contains free variables

method reify : (Env.t -> ('a, 'b) injected -> 'b) -> 'b

Gets the answer as a logic value using provided injection function inj

method prjc : (Env.t -> ('a, 'b) injected -> 'a) -> 'a