Module OCanren__.Core

module State : sig ... end

type 'a goal'

Goal converts a state into a lazy stream of states

type goal = State.t OCanren.Stream.t goal'

miniKanren basic combinators

val call_fresh : (('a, 'b) OCanren__.Logic.injected -> goal) -> goal

call_fresh f creates a fresh logical variable and passes it to the parameter

val (===) : ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> goal

x === y creates a goal, which performs a unification of x and y

val unify : ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> goal

unify x y is a prefix synonym for x === y

val (=/=) : ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> goal

x =/= y creates a goal, which introduces a disequality constraint for x and y

val structural : ('a, 'b) OCanren__.Logic.injected -> (OCanren.Env.t -> ('a, 'b) OCanren__.Logic.injected -> 'b) -> ('b -> bool) -> goal

Call structural var reifier checker adds a structural constraint for future use. Every time substitution is updated it reifies var using reifier and checks that the result satisfies desired predicate checker.

The predicate checker returns false when constraint is violated. Call structural saves constraint and return substitution nonmodified.

See also: debug_var.

val diseq : ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> ('a, 'b OCanren__.Logic.logic) OCanren__.Logic.injected -> goal

diseq x y is a prefix synonym for x =/= y

val conj : goal -> goal -> goal

conj s1 s2 creates a goal, which is a conjunction of its arguments

val (&&&) : goal -> goal -> goal

&&& is a left-associative infix synonym for conj

val disj : goal -> goal -> goal

disj s1 s2 creates a goal, which is a disjunction of its arguments

val (|||) : goal -> goal -> goal

||| is a left-associative infix synonym for disj

val (?|) : goal list -> goal

?| [s1; s2; ...; sk] calculates s1 ||| (s2 ||| ... ||| sk)...) for a non-empty list of goals (note the right association)

val conde : goal list -> goal

conde is a synonym for ?|

val (?&) : goal list -> goal

?& [s1; s2; ...; sk] calculates s1 &&& (s2 && ... &&& sk)...) for a non-empty list of goals (note the right association)

Some predefined goals

val success : goal

success always succeeds

val failure : goal

failure always fails

module Fresh : sig ... end

Top-level running primitives

val run : (unit -> ('a -> State.t -> 'b) * ('c -> OCanren.Env.t -> 'd) * ('b -> 'c * State.t OCanren.Stream.t) * ('e -> 'd -> 'f)) -> 'a -> 'e -> 'f OCanren.Stream.t

run n g h runs a goal g with n logical parameters and passes reified results to the handler h. The number of parameters is encoded using variadic machinery à la Olivier Danvy and represented by a number of predefined numerals and successor function (see below). The reification replaces each variable, passed to g, with the stream of values, associated with that variable as the goal succeeds.

Examples:

run one (fun q -> q === !!5) (fun qs -> ...). Here qs --- a stream of all values, associated with the variable q.

run two (fun q r -> q === !!5 ||| r === !!6) (fun qs rs -> ...). Here qs, rs --- streams of all values, associated with the variable q and r, respectively.

run (succ one) (fun q r -> q === !!5 ||| r === !!6) (fun qs rs -> ...). The same as the above.

val delay : (unit -> goal) -> goal

The primitive delay helps to construct recursive goals, which depend on themselves. For example, we can't write let rec fives q = (q === !!5) ||| (fives q) because the generation of this goal leads to infinite recursion. The correct way to implement this is let rec fives q = (q === !!5) ||| delay (fun () -> fives q)

See also syntax extension defer.

val succ : (unit -> ('a -> State.t -> 'b) * ('c -> OCanren.Env.t -> 'd) * ('e -> 'f * 'g) * ('h -> 'i -> 'j)) -> unit -> ((('k, 'l) OCanren__.Logic.injected -> 'a) -> State.t -> ('k, 'l) OCanren__.Logic.injected * 'b) * ((('m, 'n) OCanren__.Logic.injected * 'c) -> OCanren.Env.t -> ('m, 'n) OCanren__.Logic.reified * 'd) * (('o * 'e) -> ('o * 'f) * 'g) * (('p -> 'h) -> ('p * 'i) -> 'j)

Successor function

val one : unit -> ((('a, 'b) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * State.t OCanren.Stream.t) * (('c, 'd) OCanren__.Logic.injected -> OCanren.Env.t -> ('c, 'd) OCanren__.Logic.reified) * ('e -> 'e) * (('f -> 'g) -> 'f -> 'g)

Predefined numerals (one to five)

val two : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * State.t OCanren.Stream.t)) * ((('e, 'f) OCanren__.Logic.injected * ('g, 'h) OCanren__.Logic.injected) -> OCanren.Env.t -> ('e, 'f) OCanren__.Logic.reified * ('g, 'h) OCanren__.Logic.reified) * (('i * ('j * 'k)) -> ('i * 'j) * 'k) * (('l -> 'm -> 'n) -> ('l * 'm) -> 'n)

val three : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * State.t OCanren.Stream.t))) * ((('g, 'h) OCanren__.Logic.injected * (('i, 'j) OCanren__.Logic.injected * ('k, 'l) OCanren__.Logic.injected)) -> OCanren.Env.t -> ('g, 'h) OCanren__.Logic.reified * (('i, 'j) OCanren__.Logic.reified * ('k, 'l) OCanren__.Logic.reified)) * (('m * ('n * ('o * 'p))) -> ('m * ('n * 'o)) * 'p) * (('q -> 'r -> 's -> 't) -> ('q * ('r * 's)) -> 't)

val four : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * (('g, 'h) OCanren__.Logic.injected * State.t OCanren.Stream.t)))) * ((('i, 'j) OCanren__.Logic.injected * (('k, 'l) OCanren__.Logic.injected * (('m, 'n) OCanren__.Logic.injected * ('o, 'p) OCanren__.Logic.injected))) -> OCanren.Env.t -> ('i, 'j) OCanren__.Logic.reified * (('k, 'l) OCanren__.Logic.reified * (('m, 'n) OCanren__.Logic.reified * ('o, 'p) OCanren__.Logic.reified))) * (('q * ('r * ('s * ('t * 'u)))) -> ('q * ('r * ('s * 't))) * 'u) * (('v -> 'w -> 'x -> 'y -> 'z) -> ('v * ('w * ('x * 'y))) -> 'z)

val five : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> ('i, 'j) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * (('g, 'h) OCanren__.Logic.injected * (('i, 'j) OCanren__.Logic.injected * State.t OCanren.Stream.t))))) * ((('k, 'l) OCanren__.Logic.injected * (('m, 'n) OCanren__.Logic.injected * (('o, 'p) OCanren__.Logic.injected * (('q, 'r) OCanren__.Logic.injected * ('s, 't) OCanren__.Logic.injected)))) -> OCanren.Env.t -> ('k, 'l) OCanren__.Logic.reified * (('m, 'n) OCanren__.Logic.reified * (('o, 'p) OCanren__.Logic.reified * (('q, 'r) OCanren__.Logic.reified * ('s, 't) OCanren__.Logic.reified)))) * (('u * ('v * ('w * ('x * ('y * 'z))))) -> ('u * ('v * ('w * ('x * 'y)))) * 'z) * (('a1 -> 'b1 -> 'c1 -> 'd1 -> 'e1 -> 'f1) -> ('a1 * ('b1 * ('c1 * ('d1 * 'e1)))) -> 'f1)

val q : unit -> ((('a, 'b) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * State.t OCanren.Stream.t) * (('c, 'd) OCanren__.Logic.injected -> OCanren.Env.t -> ('c, 'd) OCanren__.Logic.reified) * ('e -> 'e) * (('f -> 'g) -> 'f -> 'g)

The same numerals with conventional names

val qr : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * State.t OCanren.Stream.t)) * ((('e, 'f) OCanren__.Logic.injected * ('g, 'h) OCanren__.Logic.injected) -> OCanren.Env.t -> ('e, 'f) OCanren__.Logic.reified * ('g, 'h) OCanren__.Logic.reified) * (('i * ('j * 'k)) -> ('i * 'j) * 'k) * (('l -> 'm -> 'n) -> ('l * 'm) -> 'n)

val qrs : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * State.t OCanren.Stream.t))) * ((('g, 'h) OCanren__.Logic.injected * (('i, 'j) OCanren__.Logic.injected * ('k, 'l) OCanren__.Logic.injected)) -> OCanren.Env.t -> ('g, 'h) OCanren__.Logic.reified * (('i, 'j) OCanren__.Logic.reified * ('k, 'l) OCanren__.Logic.reified)) * (('m * ('n * ('o * 'p))) -> ('m * ('n * 'o)) * 'p) * (('q -> 'r -> 's -> 't) -> ('q * ('r * 's)) -> 't)

val qrst : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * (('g, 'h) OCanren__.Logic.injected * State.t OCanren.Stream.t)))) * ((('i, 'j) OCanren__.Logic.injected * (('k, 'l) OCanren__.Logic.injected * (('m, 'n) OCanren__.Logic.injected * ('o, 'p) OCanren__.Logic.injected))) -> OCanren.Env.t -> ('i, 'j) OCanren__.Logic.reified * (('k, 'l) OCanren__.Logic.reified * (('m, 'n) OCanren__.Logic.reified * ('o, 'p) OCanren__.Logic.reified))) * (('q * ('r * ('s * ('t * 'u)))) -> ('q * ('r * ('s * 't))) * 'u) * (('v -> 'w -> 'x -> 'y -> 'z) -> ('v * ('w * ('x * 'y))) -> 'z)

val qrstu : unit -> ((('a, 'b) OCanren__.Logic.injected -> ('c, 'd) OCanren__.Logic.injected -> ('e, 'f) OCanren__.Logic.injected -> ('g, 'h) OCanren__.Logic.injected -> ('i, 'j) OCanren__.Logic.injected -> goal) -> State.t -> ('a, 'b) OCanren__.Logic.injected * (('c, 'd) OCanren__.Logic.injected * (('e, 'f) OCanren__.Logic.injected * (('g, 'h) OCanren__.Logic.injected * (('i, 'j) OCanren__.Logic.injected * State.t OCanren.Stream.t))))) * ((('k, 'l) OCanren__.Logic.injected * (('m, 'n) OCanren__.Logic.injected * (('o, 'p) OCanren__.Logic.injected * (('q, 'r) OCanren__.Logic.injected * ('s, 't) OCanren__.Logic.injected)))) -> OCanren.Env.t -> ('k, 'l) OCanren__.Logic.reified * (('m, 'n) OCanren__.Logic.reified * (('o, 'p) OCanren__.Logic.reified * (('q, 'r) OCanren__.Logic.reified * ('s, 't) OCanren__.Logic.reified)))) * (('u * ('v * ('w * ('x * ('y * 'z))))) -> ('u * ('v * ('w * ('x * 'y)))) * 'z) * (('a1 -> 'b1 -> 'c1 -> 'd1 -> 'e1 -> 'f1) -> ('a1 * ('b1 * ('c1 * ('d1 * 'e1)))) -> 'f1)

module Tabling : sig ... end

Tabling primitives. Tabling allows to cache answers of the goal between different queries.

val debug_var : ('a, 'b) OCanren__.Logic.injected -> (('a, 'b) OCanren__.Logic.injected -> OCanren.Env.t -> 'b) -> ('b list -> goal) -> goal

The call debug_var var reifier callback performs reification of variable var in a current state using reifier and passes list of answer to callback (multiple answers can arise in presence of disequality constraints). The callback can investigate reified value and construct required goal to continue search.

See also: structural.

val only_head : goal -> goal

The goal only_head f returns no answers when f returns:

empty stream when f returns empty stream;

hangs when f hangs during search for first answer;

stream with head answer if f returns stream that has at least 1 answer.

module PrunesControl : sig ... end