Module OCanren__.List

Relational Lists

type ('a, 'l) list =

| Nil

| Cons of 'a * 'l

Abstract list type

class virtual ['ia, 'a, 'sa, 'il, 'l, 'sl, 'inh, 'extra, 'syn] list_t : object ... end

val gcata_list : < c_Nil : 'inh -> ('a, 'l) list -> 'syn; c_Cons : 'inh -> ('a, 'l) list -> 'a -> 'l -> 'syn; .. > -> 'inh -> ('a, 'l) list -> 'syn

class ['a, 'l, 'extra_list] fmt_list_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> (Stdlib.Format.formatter -> ('a, 'l) list -> unit) -> object ... end

val fmt_list : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> Stdlib.Format.formatter -> ('a, 'l) list -> unit

class ['a, 'l, 'syn, 'extra_list] foldr_list_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> ('syn -> ('a, 'l) list -> 'syn) -> object ... end

val foldr_list : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) list -> 'syn

class ['a, 'l, 'syn, 'extra_list] foldl_list_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> ('syn -> ('a, 'l) list -> 'syn) -> object ... end

val foldl_list : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) list -> 'syn

class ['a, 'l, 'extra_list] compare_list_t : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> (('a, 'l) list -> ('a, 'l) list -> GT.comparison) -> object ... end

val compare_list : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> ('a, 'l) list -> ('a, 'l) list -> GT.comparison

class ['a, 'l, 'extra_list] eq_list_t : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> (('a, 'l) list -> ('a, 'l) list -> bool) -> object ... end

val eq_list : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> ('a, 'l) list -> ('a, 'l) list -> bool

class ['a, 'l, 'extra_list] html_list_t : (unit -> 'a -> HTML.er) -> (unit -> 'l -> HTML.er) -> (unit -> ('a, 'l) list -> HTML.er) -> object ... end

val html_list : ('a -> HTML.er) -> ('l -> HTML.er) -> ('a, 'l) list -> HTML.er

class ['a, 'a_2, 'l, 'l_2, 'extra_list, 'syn_list] gmap_list_t : (unit -> 'a -> 'a_2) -> (unit -> 'l -> 'l_2) -> (unit -> ('a, 'l) list -> ('a_2, 'l_2) list) -> object ... end

val gmap_list : ('a -> 'a_2) -> ('l -> 'l_2) -> ('a, 'l) list -> ('a_2, 'l_2) list

class ['a, 'l, 'extra_list] show_list_t : (unit -> 'a -> string) -> (unit -> 'l -> string) -> (unit -> ('a, 'l) list -> string) -> object ... end

val show_list : ('a -> string) -> ('l -> string) -> ('a, 'l) list -> string

val list : (< c_Nil : 'inh -> ('a, 'l) list -> 'syn; c_Cons : 'inh -> ('a, 'l) list -> 'a -> 'l -> 'syn; .. > -> 'inh -> ('a, 'l) list -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> Stdlib.Format.formatter -> ('a, 'l) list -> unit; foldr : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) list -> 'syn; foldl : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) list -> 'syn; compare : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> ('a, 'l) list -> ('a, 'l) list -> GT.comparison; eq : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> ('a, 'l) list -> ('a, 'l) list -> bool; html : ('a -> HTML.er) -> ('l -> HTML.er) -> ('a, 'l) list -> HTML.er; gmap : ('a -> 'a_2) -> ('l -> 'l_2) -> ('a, 'l) list -> ('a_2, 'l_2) list; show : ('a -> string) -> ('l -> string) -> ('a, 'l) list -> string; >, (('inh3 -> ('a1, 'a2) list -> 'syn4) -> ['a1_i, 'a1, 'a1_s, 'a2_i, 'a2, 'a2_s, 'inh3, ('a1, 'a2) list, 'syn4] list_t) -> 'inh3 -> ('a1, 'a2) list -> 'syn4) GT.t

GT-related API

type 'a logic' = 'a OCanren__.Logic.logic

Type synonym to prevent toplevel logic from being hidden

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] logic'_t : object ... end

val gcata_logic' : [_, 'a, 'sa, 'inh, 'a logic', 'syn] logic'_t -> 'inh -> 'a logic' -> 'syn

class ['a, 'extra_logic'] fmt_logic'_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'a logic' -> unit) -> object ... end

val fmt_logic' : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic' -> unit

class ['a, 'syn, 'extra_logic'] foldr_logic'_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic' -> 'syn) -> object ... end

val foldr_logic' : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic' -> 'syn

class ['a, 'syn, 'extra_logic'] foldl_logic'_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic' -> 'syn) -> object ... end

val foldl_logic' : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic' -> 'syn

class ['a, 'extra_logic'] compare_logic'_t : ('a -> 'a -> GT.comparison) -> ('a logic' -> 'a logic' -> GT.comparison) -> object ... end

val compare_logic' : ('a -> 'a -> GT.comparison) -> 'a logic' -> 'a logic' -> GT.comparison

class ['a, 'extra_logic'] eq_logic'_t : ('a -> 'a -> bool) -> ('a logic' -> 'a logic' -> bool) -> object ... end

val eq_logic' : ('a -> 'a -> bool) -> 'a logic' -> 'a logic' -> bool

class ['a, 'extra_logic'] html_logic'_t : (unit -> 'a -> HTML.er) -> (unit -> 'a logic' -> HTML.er) -> object ... end

val html_logic' : ('a -> HTML.er) -> 'a logic' -> HTML.er

class ['a, 'a_2, 'extra_logic', 'syn_logic'] gmap_logic'_t : (unit -> 'a -> 'a_2) -> (unit -> 'a logic' -> 'a_2 logic') -> object ... end

val gmap_logic' : ('a -> 'a_2) -> 'a logic' -> 'a_2 logic'

class ['a, 'extra_logic'] show_logic'_t : (unit -> 'a -> string) -> (unit -> 'a logic' -> string) -> object ... end

val show_logic' : ('a -> string) -> 'a logic' -> string

val logic' : ([_, 'a, 'sa, 'inh, 'a logic', 'syn] logic'_t -> 'inh -> 'a logic' -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic' -> unit; foldr : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic' -> 'syn; foldl : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic' -> 'syn; compare : ('a -> 'a -> GT.comparison) -> 'a logic' -> 'a logic' -> GT.comparison; eq : ('a -> 'a -> bool) -> 'a logic' -> 'a logic' -> bool; html : ('a -> HTML.er) -> 'a logic' -> HTML.er; gmap : ('a -> 'a_2) -> 'a logic' -> 'a_2 logic'; show : ('a -> string) -> 'a logic' -> string; >, (('inh2 -> 'a1 logic' -> 'syn3) -> ['a1_i, 'a1, 'a1_s, 'inh2, 'a1 logic', 'syn3] logic'_t) -> 'inh2 -> 'a1 logic' -> 'syn3) GT.t

type ('a, 'l) t = ('a, 'l) list

Synonym for abstract list type

class virtual ['ia, 'a, 'sa, 'il, 'l, 'sl, 'inh, 'extra, 'syn] t_t : object ... end

val gcata_t : [_, 'a, 'sa, _, 'l, 'sl, 'inh, ('a, 'l) t, 'syn] t_t -> 'inh -> ('a, 'l) t -> 'syn

class ['a, 'l, 'extra_t] fmt_t_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> (Stdlib.Format.formatter -> ('a, 'l) t -> unit) -> object ... end

val fmt_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> Stdlib.Format.formatter -> ('a, 'l) t -> unit

class ['a, 'l, 'syn, 'extra_t] foldr_t_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> ('syn -> ('a, 'l) t -> 'syn) -> object ... end

val foldr_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) t -> 'syn

class ['a, 'l, 'syn, 'extra_t] foldl_t_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> ('syn -> ('a, 'l) t -> 'syn) -> object ... end

val foldl_t : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) t -> 'syn

class ['a, 'l, 'extra_t] compare_t_t : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> (('a, 'l) t -> ('a, 'l) t -> GT.comparison) -> object ... end

val compare_t : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> ('a, 'l) t -> ('a, 'l) t -> GT.comparison

class ['a, 'l, 'extra_t] eq_t_t : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> (('a, 'l) t -> ('a, 'l) t -> bool) -> object ... end

val eq_t : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> ('a, 'l) t -> ('a, 'l) t -> bool

class ['a, 'l, 'extra_t] html_t_t : (unit -> 'a -> HTML.er) -> (unit -> 'l -> HTML.er) -> (unit -> ('a, 'l) t -> HTML.er) -> object ... end

val html_t : ('a -> HTML.er) -> ('l -> HTML.er) -> ('a, 'l) t -> HTML.er

class ['a, 'a_2, 'l, 'l_2, 'extra_t, 'syn_t] gmap_t_t : (unit -> 'a -> 'a_2) -> (unit -> 'l -> 'l_2) -> (unit -> ('a, 'l) t -> ('a_2, 'l_2) t) -> object ... end

val gmap_t : ('a -> 'a_2) -> ('l -> 'l_2) -> ('a, 'l) t -> ('a_2, 'l_2) t

class ['a, 'l, 'extra_t] show_t_t : (unit -> 'a -> string) -> (unit -> 'l -> string) -> (unit -> ('a, 'l) t -> string) -> object ... end

val show_t : ('a -> string) -> ('l -> string) -> ('a, 'l) t -> string

val t : ([_, 'a, 'sa, _, 'l, 'sl, 'inh, ('a, 'l) t, 'syn] t_t -> 'inh -> ('a, 'l) t -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'l -> unit) -> Stdlib.Format.formatter -> ('a, 'l) t -> unit; foldr : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) t -> 'syn; foldl : ('syn -> 'a -> 'syn) -> ('syn -> 'l -> 'syn) -> 'syn -> ('a, 'l) t -> 'syn; compare : ('a -> 'a -> GT.comparison) -> ('l -> 'l -> GT.comparison) -> ('a, 'l) t -> ('a, 'l) t -> GT.comparison; eq : ('a -> 'a -> bool) -> ('l -> 'l -> bool) -> ('a, 'l) t -> ('a, 'l) t -> bool; html : ('a -> HTML.er) -> ('l -> HTML.er) -> ('a, 'l) t -> HTML.er; gmap : ('a -> 'a_2) -> ('l -> 'l_2) -> ('a, 'l) t -> ('a_2, 'l_2) t; show : ('a -> string) -> ('l -> string) -> ('a, 'l) t -> string; >, (('inh3 -> ('a1, 'a2) t -> 'syn4) -> ['a1_i, 'a1, 'a1_s, 'a2_i, 'a2, 'a2_s, 'inh3, ('a1, 'a2) t, 'syn4] t_t) -> 'inh3 -> ('a1, 'a2) t -> 'syn4) GT.t

type 'a ground = ('a, 'a ground) t

Ground lists (isomorphic to regular ones)

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] ground_t : object ... end

val gcata_ground : [_, 'a, 'sa, 'inh, 'a ground, 'syn] ground_t -> 'inh -> 'a ground -> 'syn

class ['a, 'extra_ground] fmt_ground_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'a ground -> unit) -> object ... end

val fmt_ground : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a ground -> unit

class ['a, 'syn, 'extra_ground] foldr_ground_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a ground -> 'syn) -> object ... end

val foldr_ground : ('syn -> 'a -> 'syn) -> 'syn -> 'a ground -> 'syn

class ['a, 'syn, 'extra_ground] foldl_ground_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a ground -> 'syn) -> object ... end

val foldl_ground : ('syn -> 'a -> 'syn) -> 'syn -> 'a ground -> 'syn

class ['a, 'extra_ground] compare_ground_t : ('a -> 'a -> GT.comparison) -> ('a ground -> 'a ground -> GT.comparison) -> object ... end

val compare_ground : ('a -> 'a -> GT.comparison) -> 'a ground -> 'a ground -> GT.comparison

class ['a, 'extra_ground] eq_ground_t : ('a -> 'a -> bool) -> ('a ground -> 'a ground -> bool) -> object ... end

val eq_ground : ('a -> 'a -> bool) -> 'a ground -> 'a ground -> bool

class ['a, 'extra_ground] html_ground_t : (unit -> 'a -> HTML.er) -> (unit -> 'a ground -> HTML.er) -> object ... end

val html_ground : ('a -> HTML.er) -> 'a ground -> HTML.er

class ['a, 'a_2, 'extra_ground, 'syn_ground] gmap_ground_t : (unit -> 'a -> 'a_2) -> (unit -> 'a ground -> 'a_2 ground) -> object ... end

val gmap_ground : ('a -> 'a_2) -> 'a ground -> 'a_2 ground

class ['a, 'extra_ground] show_ground_t : (unit -> 'a -> string) -> (unit -> 'a ground -> string) -> object ... end

val show_ground : ('a -> string) -> 'a ground -> string

val ground : ([_, 'a, 'sa, 'inh, 'a ground, 'syn] ground_t -> 'inh -> 'a ground -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a ground -> unit; foldr : ('syn -> 'a -> 'syn) -> 'syn -> 'a ground -> 'syn; foldl : ('syn -> 'a -> 'syn) -> 'syn -> 'a ground -> 'syn; compare : ('a -> 'a -> GT.comparison) -> 'a ground -> 'a ground -> GT.comparison; eq : ('a -> 'a -> bool) -> 'a ground -> 'a ground -> bool; html : ('a -> HTML.er) -> 'a ground -> HTML.er; gmap : ('a -> 'a_2) -> 'a ground -> 'a_2 ground; show : ('a -> string) -> 'a ground -> string; >, (('inh2 -> 'a1 ground -> 'syn3) -> ['a1_i, 'a1, 'a1_s, 'inh2, 'a1 ground, 'syn3] ground_t) -> 'inh2 -> 'a1 ground -> 'syn3) GT.t

type 'a logic = ('a, 'a logic) t logic'

Logic lists (with the tails as logic lists)

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] logic_t : object ... end

val gcata_logic : [_, 'a, 'sa, 'inh, 'a logic, 'syn] logic_t -> 'inh -> 'a logic -> 'syn

class ['a, 'extra_logic] fmt_logic_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'a logic -> unit) -> object ... end

val fmt_logic : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic -> unit

class ['a, 'syn, 'extra_logic] foldr_logic_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic -> 'syn) -> object ... end

val foldr_logic : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn

class ['a, 'syn, 'extra_logic] foldl_logic_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic -> 'syn) -> object ... end

val foldl_logic : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn

class ['a, 'extra_logic] compare_logic_t : ('a -> 'a -> GT.comparison) -> ('a logic -> 'a logic -> GT.comparison) -> object ... end

val compare_logic : ('a -> 'a -> GT.comparison) -> 'a logic -> 'a logic -> GT.comparison

class ['a, 'extra_logic] eq_logic_t : ('a -> 'a -> bool) -> ('a logic -> 'a logic -> bool) -> object ... end

val eq_logic : ('a -> 'a -> bool) -> 'a logic -> 'a logic -> bool

class ['a, 'extra_logic] html_logic_t : (unit -> 'a -> HTML.er) -> (unit -> 'a logic -> HTML.er) -> object ... end

val html_logic : ('a -> HTML.er) -> 'a logic -> HTML.er

class ['a, 'a_2, 'extra_logic, 'syn_logic] gmap_logic_t : (unit -> 'a -> 'a_2) -> (unit -> 'a logic -> 'a_2 logic) -> object ... end

val gmap_logic : ('a -> 'a_2) -> 'a logic -> 'a_2 logic

class ['a, 'extra_logic] show_logic_t : (unit -> 'a -> string) -> (unit -> 'a logic -> string) -> object ... end

val show_logic : ('a -> string) -> 'a logic -> string

val logic : ([_, 'a, 'sa, 'inh, 'a logic, 'syn] logic_t -> 'inh -> 'a logic -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic -> unit; foldr : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn; foldl : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn; compare : ('a -> 'a -> GT.comparison) -> 'a logic -> 'a logic -> GT.comparison; eq : ('a -> 'a -> bool) -> 'a logic -> 'a logic -> bool; html : ('a -> HTML.er) -> 'a logic -> HTML.er; gmap : ('a -> 'a_2) -> 'a logic -> 'a_2 logic; show : ('a -> string) -> 'a logic -> string; >, (('inh2 -> 'a1 logic -> 'syn3) -> ['a1_i, 'a1, 'a1_s, 'inh2, 'a1 logic, 'syn3] logic_t) -> 'inh2 -> 'a1 logic -> 'syn3) GT.t

Relational API

type ('a, 'b) groundi = ('a ground, 'b logic) OCanren__.Logic.injected

A synonym for injected list

Constructors

val nil : unit -> ('a, 'b) groundi

val cons : ('a, 'b) OCanren__.Logic.injected -> ('a, 'b) groundi -> ('a, 'b) groundi

val (%) : ('a, 'b) OCanren__.Logic.injected -> ('a, 'b) groundi -> ('a, 'b) groundi

Infix synonym for cons

val (%<) : ('a, 'b) OCanren__.Logic.injected -> ('a, 'b) OCanren__.Logic.injected -> ('a, 'b) groundi

x %< y is a synonym for cons x (cons y (nil ()))

val (!<) : ('a, 'b) OCanren__.Logic.injected -> ('a, 'b) groundi

!< x is a synonym for cons x (nil ())

Built-in relations

val of_list : ('a -> 'b) -> 'a GT.list -> 'b ground

of_list l converts regular OCaml list l into isomorphic OCanren ground list

val to_list : ('a -> 'b) -> 'a ground -> 'b GT.list

to_list g converts OCanren list g into regular OCaml list

val inj : ('a -> 'b) -> 'a ground -> 'b logic

inj x makes a logic list from a ground one

val list : ('a, 'b) OCanren__.Logic.injected GT.list -> ('a, 'b) groundi

Make injected list from ground one of injected elements

val reify : (OCanren.Env.t -> ('a, 'b) OCanren__.Logic.injected -> 'b) -> OCanren.Env.t -> ('a ground, 'b logic) OCanren__.Logic.injected -> 'b logic

Reifier

val prjc : (OCanren.Env.t -> ('a, 'b) OCanren__.Logic.injected -> 'a) -> (int -> 'a ground GT.list -> 'a ground) -> OCanren.Env.t -> ('a ground, 'b logic) OCanren__.Logic.injected -> 'a ground

val foldro : (('a, 'b) OCanren__.Logic.injected -> ('acc, _ logic' as 'acc2) OCanren__.Logic.injected -> ('acc, 'acc2) OCanren__.Logic.injected -> OCanren__.Core.goal) -> ('acc, 'acc2) OCanren__.Logic.injected -> ('a, 'b) groundi -> ('acc, 'acc2) OCanren__.Logic.injected -> OCanren__.Core.goal

Relational foldr

val mapo : (('a, 'b) OCanren__.Logic.injected -> ('q, 'w) OCanren__.Logic.injected -> OCanren__.Core.goal) -> ('a, 'b) groundi -> ('q, 'w) groundi -> OCanren__.Core.goal

Relational map

val filtero : (('a, 'b) OCanren__.Logic.injected -> OCanren__.Bool.groundi -> OCanren__.Core.goal) -> ('a, 'b) groundi -> ('a, 'b) groundi -> OCanren__.Core.goal

Relational filter

val lookupo : (('a, 'b) OCanren__.Logic.injected -> OCanren__.Bool.groundi -> OCanren__.Core.goal) -> ('a, 'b) groundi -> ('a option, 'b option logic') OCanren__.Logic.injected -> OCanren__.Core.goal

Relational lookup

val assoco : ('a, 'b logic') OCanren__.Logic.injected -> (('a, 'c) OCanren__.Pair.ground, ('b logic', 'd logic') OCanren__.Pair.logic) groundi -> ('c, 'd logic') OCanren__.Logic.injected -> OCanren__.Core.goal

Relational association list lookup

val anyo : (OCanren__.Bool.ground, OCanren__.Bool.logic) groundi -> OCanren__.Bool.groundi -> OCanren__.Core.goal

Boolean list disjunctions

val allo : (OCanren__.Bool.ground, OCanren__.Bool.logic) groundi -> OCanren__.Bool.groundi -> OCanren__.Core.goal

Boolean list conjunction

val lengtho : (_, _) groundi -> OCanren__.Nat.groundi -> OCanren__.Core.goal

Relational length

val appendo : ('a, 'b) groundi -> ('a, 'b) groundi -> ('a, 'b) groundi -> OCanren__.Core.goal

Relational append

val reverso : ('a, 'b) groundi -> ('a, 'b) groundi -> OCanren__.Core.goal

Relational reverse

val membero : ('a, 'b logic') groundi -> ('a, 'b logic') OCanren__.Logic.injected -> OCanren__.Core.goal

Relational occurrence check (a shortcut)

val nullo : (_, _) groundi -> OCanren__.Core.goal

Relational check for empty list

val caro : ('a, 'b) groundi -> ('a, 'b) OCanren__.Logic.injected -> OCanren__.Core.goal

Relational head of the list

val hdo : ('a, 'b) groundi -> ('a, 'b) OCanren__.Logic.injected -> OCanren__.Core.goal

Alias for caro

val cdro : ('a, 'b) groundi -> ('a, 'b) groundi -> OCanren__.Core.goal

Relational tail of the list

val tlo : ('a, 'b) groundi -> ('a, 'b) groundi -> OCanren__.Core.goal

Alias for cdro