Module OCanren__Logic

Logic values

type 'a logic = private

| Var of GT.int * 'a logic GT.list

| Value of 'a

A type of a logic value

class virtual ['ia, 'a, 'sa, 'inh, 'extra, 'syn] logic_t : object ... end

val gcata_logic : < c_Var : 'inh -> 'a logic -> GT.int -> 'a logic GT.list -> 'syn; c_Value : 'inh -> 'a logic -> 'a -> 'syn; .. > -> 'inh -> 'a logic -> 'syn

class ['a, 'extra_logic] fmt_logic_t : (Stdlib.Format.formatter -> 'a -> unit) -> (Stdlib.Format.formatter -> 'a logic -> unit) -> object ... end

val fmt_logic : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic -> unit

class ['a, 'syn, 'extra_logic] foldr_logic_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic -> 'syn) -> object ... end

val foldr_logic : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn

class ['a, 'syn, 'extra_logic] foldl_logic_t : ('syn -> 'a -> 'syn) -> ('syn -> 'a logic -> 'syn) -> object ... end

val foldl_logic : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn

class ['a, 'extra_logic] compare_logic_t : ('a -> 'a -> GT.comparison) -> ('a logic -> 'a logic -> GT.comparison) -> object ... end

val compare_logic : ('a -> 'a -> GT.comparison) -> 'a logic -> 'a logic -> GT.comparison

class ['a, 'extra_logic] eq_logic_t : ('a -> 'a -> bool) -> ('a logic -> 'a logic -> bool) -> object ... end

val eq_logic : ('a -> 'a -> bool) -> 'a logic -> 'a logic -> bool

class ['a, 'extra_logic] html_logic_t : (unit -> 'a -> HTML.er) -> (unit -> 'a logic -> HTML.er) -> object ... end

val html_logic : ('a -> HTML.er) -> 'a logic -> HTML.er

class ['a, 'a_2, 'extra_logic, 'syn_logic] gmap_logic_t : (unit -> 'a -> 'a_2) -> (unit -> 'a logic -> 'a_2 logic) -> object ... end

val gmap_logic : ('a -> 'a_2) -> 'a logic -> 'a_2 logic

class ['a, 'extra_logic] show_logic_t : (unit -> 'a -> string) -> (unit -> 'a logic -> string) -> object ... end

val show_logic : ('a -> string) -> 'a logic -> string

val logic : (< c_Var : 'inh -> 'a logic -> GT.int -> 'a logic GT.list -> 'syn; c_Value : 'inh -> 'a logic -> 'a -> 'syn; .. > -> 'inh -> 'a logic -> 'syn, < fmt : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a logic -> unit; foldr : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn; foldl : ('syn -> 'a -> 'syn) -> 'syn -> 'a logic -> 'syn; compare : ('a -> 'a -> GT.comparison) -> 'a logic -> 'a logic -> GT.comparison; eq : ('a -> 'a -> bool) -> 'a logic -> 'a logic -> bool; html : ('a -> HTML.er) -> 'a logic -> HTML.er; gmap : ('a -> 'a_2) -> 'a logic -> 'a_2 logic; show : ('a -> string) -> 'a logic -> string; >, (('inh2 -> 'a1 logic -> 'syn3) -> ['a1_i, 'a1, 'a1_s, 'inh2, 'a1 logic, 'syn3] logic_t) -> 'inh2 -> 'a1 logic -> 'syn3) GT.t

val to_logic : 'a -> 'a logic

to_logic x makes a logic value from a regular one

val from_logic : 'a logic -> 'a

from_logic x makes a regular value from a logic one. Raises exception Not_a_value if x contains free variables

Injections/projections

type ('a, 'b) injected

The type ('a, 'b) injected describes an injection of a type 'a into 'b

val lift : 'a -> ('a, 'a) injected

lift x lifts x into injected doamin

val inj : ('a, 'b) injected -> ('a, 'b logic) injected

inj x injects x into logical x

val (!!) : 'a -> ('a, 'a logic) injected

A synonym for fun x -> inj @@ lift x (for non-parametric types)

val prj : ('a, 'b) injected -> 'a

prj x returns a regular value from injected representation. Raises exception Not_a_value if x contains free variables

exception Not_a_value

The exception is raised when we try to extract a regular term from the answer with some free variables

class type ['a, 'b] reified = object ... end

Reification result

val make_rr : OCanren.Env.t -> ('a, 'b) injected -> ('a, 'b) reified

module type T1 = sig ... end

module type T2 = sig ... end

module type T3 = sig ... end

module type T4 = sig ... end

module type T5 = sig ... end

module type T6 = sig ... end

val reify : OCanren.Env.t -> ('a, 'a logic) injected -> 'a logic

val prjc : (int -> 'a list -> 'a) -> OCanren.Env.t -> ('a, 'a logic) injected -> 'a

val project : ('a, 'b) reified -> 'a

module Fmap : functor (T : T1) -> sig ... end

module Fmap2 : functor (T : T2) -> sig ... end

module Fmap3 : functor (T : T3) -> sig ... end

module Fmap4 : functor (T : T4) -> sig ... end

module Fmap5 : functor (T : T5) -> sig ... end

module Fmap6 : functor (T : T6) -> sig ... end